wfc_blender_hash/pyproject.toml
[project]
name = "wfc_blender_hash"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Hari Krishnan", email = "codeharik@gmail.com" }
]
dependencies = [
    "fake-bpy-module-latest>=20250630",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
typeCheckingMode = "basic"
reportInvalidTypeForm = false
reportAttributeAccessIssue = false
reportOptionalMemberAccess = false
reportArgumentType = false
reportSelfClsParameterName = false
